bot/requirements.txt
python-telegram-bot==21.6
requests==2.32.3
pydantic==2.10.3
pydantic-settings==2.6.1
